Removed profileinfosection component
